refactor(ecommerce): small refactorings
- move AuthService.kt
- CartItemDto.kt does not need userID
- rename id to cart_id in schema.sql
- TokenLoginController.kt becomes a RestController
