+++ src/main/kotlin/ecommerce/dto/CartItemDto.kt
@@ -2,7 +2,6 @@ package ecommerce.dto
 
 class CartItemDto(
     val cartId: Long,
-    val userId: Long,
     val productId: Long,
     val quantity: Int,
 )
